+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Add functionality to refresh bearer token
 - Add examples  [ `authenticate_user.rs` , `refresh_token.rs`, `revoke_token.rs` ]
+- Add Listing and Post models
 
 ## 0.1.0 (2020/10/07)